+++ test/unittest/verifierSpec.mjs.js
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020-2024 Philippe Lieser
+ * Copyright (c) 2020-2025 Philippe Lieser
  *
  * This software is licensed under the terms of the MIT License.
  *
@@ -195,6 +195,15 @@ describe("DKIM Verifier [unittest]", function () {
 			expect(res.signatures[0]?.warnings).to.be.empty;
 		});
 
+		it("DKIM key with 's' flag - AUID and SDID match", async function () {
+			const res = await verifyEmlFile("dkim/flags_s-auid_and_sdid_match.eml");
+			expect(res.signatures.length).to.be.equal(1);
+			expect(res.signatures[0]?.result).to.be.equal("SUCCESS");
+			expect(res.signatures[0]?.warnings).to.be.empty;
+			expect(res.signatures[0]?.sdid).to.be.equal("example.com");
+			expect(res.signatures[0]?.auid).to.be.equal("@example.com");
+		});
+
 		it("Received time is after Signature Timestamp", async function () {
 			const res = await verifyEmlFile("dkim/time-received_after_creation.eml");
 			expect(res.signatures.length).to.be.equal(1);
@@ -268,18 +277,13 @@ describe("DKIM Verifier [unittest]", function () {
 			expect(res.signatures[0]?.errorType).to.be.equal("DKIM_SIGERROR_KEY_NOTEMAILKEY");
 		});
 
-		it("DKIM key does not allow AUID to be a subdomain", async function () {
-			const res = await verifyEmlFile("rfc6376-A.2.eml", new Map([
-				["brisbane._domainkey.example.com",
-					"v=DKIM1;t=s;p=" +
-					"MIGfMA0GCSqGSIb3DQEBAQUAA4GNADCBiQKBgQDwIRP/UC3SBsEmGqZ9ZJW3/DkM" +
-					"oGeLnQg1fWn7/zYtIxN2SnFCjxOCKG9v3b4jYfcTNh5ijSsq631uBItLa7od+v/R" +
-					"tdC2UzJ1lWT947qR+Rcac2gbto/NMqJ0fzfVjH4OuKhitdY9tf6mcwGjaNBcWToI" +
-					"MmPSPDdQPNUYckcQ2QIDAQAB"],
-			]));
+		it("DKIM key with 's' flag  - AUID is subdomain of SDID", async function () {
+			const res = await verifyEmlFile("dkim/flags_s-auid_is_subdomain_of_sdid.eml");
 			expect(res.signatures.length).to.be.equal(1);
 			expect(res.signatures[0]?.result).to.be.equal("PERMFAIL");
 			expect(res.signatures[0]?.errorType).to.be.equal("DKIM_SIGERROR_DOMAIN_I");
+			expect(res.signatures[0]?.sdid).to.be.equal("example.com");
+			expect(res.signatures[0]?.auid).to.be.equal("@football.example.com");
 		});
 	});
 
@@ -310,24 +314,49 @@ describe("DKIM Verifier [unittest]", function () {
 		});
 	});
 
-	describe("Signature warnings", function () {
-		it("From address is not in the SDID ", async function () {
-			// TODO: instead of artificial test, add test mail there this is the case
-			const msgPlain = await readTestFile("rfc6376-A.2.eml");
-			const msgParsed = MsgParser.parseMsg(msgPlain);
-			const msg = {
-				headerFields: msgParsed.headers,
-				bodyPlain: msgParsed.body,
-				from: "foo@bar.com",
-			};
-			const verifier = new Verifier(new KeyStore(queryDnsTxt));
-			const res = await verifier.verify(msg);
+	describe("From alignment", function () {
+		it("From domain is the same as the SDID and AUID", async function () {
+			// From: joe@football.example.com
+			const res = await verifyEmlFile("rfc8463-A.3.eml");
+			expect(res.signatures[0]?.sdid).to.be.equal("football.example.com");
+			expect(res.signatures[0]?.auid).to.be.equal("@football.example.com");
+			expect(res.signatures[0]?.result).to.be.equal("SUCCESS");
+			expect(res.signatures[0]?.warnings).to.be.empty;
+		});
+
+		it("From domain is in the SDID and in the AUID", async function () {
+			// From: joe@football.example.com
+			const res = await verifyEmlFile("dkim/alignment-signed_by_parent_domain.eml");
 			expect(res.signatures.length).to.be.equal(1);
+			expect(res.signatures[0]?.sdid).to.be.equal("example.com");
+			expect(res.signatures[0]?.auid).to.be.equal("@example.com");
+			expect(res.signatures[0]?.result).to.be.equal("SUCCESS");
+			expect(res.signatures[0]?.warnings).to.be.empty;
+		});
+
+		it("From domain is in the SDID but not in the AUID", async function () {
+			// From: joe@football.example.com
+			const res = await verifyEmlFile("dkim/alignment-from_not_in_auid.eml");
+			expect(res.signatures.length).to.be.equal(1);
+			expect(res.signatures[0]?.sdid).to.be.equal("example.com");
+			expect(res.signatures[0]?.auid).to.be.equal("@support.example.com");
+			expect(res.signatures[0]?.result).to.be.equal("SUCCESS");
+			expect(res.signatures[0]?.warnings).to.be.empty;
+		});
+
+		it("From domain is not in the SDID", async function () {
+			// From: joe@football.example.net
+			const res = await verifyEmlFile("dkim/alignment-from_not_in_sdid.eml");
+			expect(res.signatures.length).to.be.equal(1);
+			expect(res.signatures[0]?.sdid).to.be.equal("example.com");
+			expect(res.signatures[0]?.auid).to.be.equal("@example.com");
 			expect(res.signatures[0]?.result).to.be.equal("SUCCESS");
 			expect(res.signatures[0]?.warnings).to.be.an("array").
 				that.deep.includes({ name: "DKIM_SIGWARNING_FROM_NOT_IN_SDID" });
 		});
+	});
 
+	describe("Signature warnings", function () {
 		it("Received time is before Signature Timestamp", async function () {
 			const res = await verifyEmlFile("dkim/time-received_long_before_creation.eml");
 			expect(res.signatures.length).to.be.equal(1);